# specs/001-agentic-url-shortener/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Agentic URL Shortener API
  version: 1.0.0
paths:
...
  /api/v1/workflows:
    post:
      summary: Create a governed SDLC workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requirement]
              properties:
                requirement: {type: string}
                scenario: {type: string, enum: [GREENFIELD, BROWNFIELD, AMBIGUOUS]}
      responses:
        '202': {description: Workflow accepted}
...
  /api/v1/workflows/{id}/clarification:
    post:
      summary: Record an explicit human clarification decision
      parameters:
        - {in: path, name: id, required: true, schema: {type: string, format: uuid}}
        - {in: query, name: decision, required: true, schema: {type: string, enum: [APPROVE, REJECT]}}
        - {in: query, name: approver, required: true, schema: {type: string}}
        - {in: query, name: comment, required: false, schema: {type: string}}
      responses:
        '200': {description: Clarification decision recorded}
...